+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # OpenCore Legacy Patcher changelog
 
 ## 1.1.0
+- Add error handling for corrupted patcher settings
 
 ## 1.0.0
 - Resolve BCM2046 and BCM2070 support on macOS 13.3 and newer